updated functions
i updated the functions which create the inner cells. I overcame the issue i had last night creating the new elements. i was missing a return statement in a function.

i went on to make the input field function wich goes on to add classes in the fields created and can if required disable the input field,